Before:
```
| Method |    Mean |   Error |   StdDev | Rank |       Gen 0 |       Gen 1 |      Gen 2 | Allocated |
|------- |--------:|--------:|---------:|-----:|------------:|------------:|-----------:|----------:|
| Format | 6.588 s | 1.049 s | 0.0575 s |    1 | 606000.0000 | 160000.0000 | 12000.0000 |   2.76 GB |
```

After:
```
| Method |    Mean |   Error |   StdDev | Rank |       Gen 0 |       Gen 1 |      Gen 2 | Allocated |
|------- |--------:|--------:|---------:|-----:|------------:|------------:|-----------:|----------:|
| Format | 6.132 s | 1.095 s | 0.0600 s |    1 | 626000.0000 | 180000.0000 | 28000.0000 |   2.76 GB |
```